llm_guardrail/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
groq>=0.4.0
python-dotenv>=0.19.0
pytest>=7.0.0  # for testing
